added exceptions for category

+++ src/main/java/com/example/practice/service/CategoryService.java
@@ -1,6 +1,10 @@
 package com.example.practice.service;
 
 import com.example.practice.entity.Category;
+import com.example.practice.exceptions.CategoryExistsException;
+import com.example.practice.exceptions.CategoryNotFoundException;
+import com.example.practice.exceptions.UserExistsException;
+import com.example.practice.exceptions.UserNotFoundException;
 import com.example.practice.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,15 +18,31 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
 
+    public void create(Category category) {
+        if (categoryRepository.existsById(category.getId())) {
+            throw new CategoryExistsException("Category with id = " + category.getId() + "already exists");
+        }
+        categoryRepository.save(category);
+    }
+
     public Optional<Category> read(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new CategoryNotFoundException();
+        }
         return categoryRepository.findById((id));
     }
 
-    public void create(Category category) {
+    public void update(Category category) {
+        if (!categoryRepository.existsById(category.getId())) {
+            throw new CategoryNotFoundException();
+        }
         categoryRepository.save(category);
     }
 
     public void delete(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new CategoryNotFoundException();
+        }
         categoryRepository.deleteById(id);
     }
 
@@ -30,12 +50,6 @@ public class CategoryService {
         return categoryRepository.findAll();
     }
 
-    public boolean update(Category category) {
-        if (categoryRepository.existsById(category.getId())) {
-            categoryRepository.save(category);
-            return true;
-        }
-        return false;
-    }
+
 
 }
